# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QA Test Case Manager API
servers:
  - url: /api
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: Invalid credentials
...
  /auth/access-requests:
    post:
      operationId: requestAccess
      tags: [auth]
      summary: Submit a pending access request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessRequestInput"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessRequest"
        "409":
          description: Duplicate email
...
  /modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List modules with test result counts
      responses:
        "200":
          description: Modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [modules]
      summary: Create an active module (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModuleInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Module" }
  /modules/{id}:
    patch:
      operationId: updateModule
      tags: [modules]
      summary: Edit or change module status (admin only)
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModuleUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Module" }
  /test-cases:
    get:
      operationId: listTestCases
      tags: [test-cases]
      summary: Search and filter test cases
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/ModuleId"
        - $ref: "#/components/parameters/Result"
        - $ref: "#/components/parameters/PerformedBy"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Test cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCasePage"
    post:
      operationId: createTestCase
      tags: [test-cases]
      summary: Create a test case with a server-generated number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestCaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCase"
  /admin/test-cases/by-number/{number}:
    get:
      operationId: getAdminTestCaseByNumber
      tags: [test-cases]
      summary: Find one test case by exact number (admin only)
      parameters:
        - name: number
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Test case
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TestCaseDetail" }
  /test-cases/{id}:
    get:
      operationId: getTestCase
      tags: [test-cases]
      summary: Get test case detail and history
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCaseDetail"
    patch:
      operationId: updateTestCase
      tags: [test-cases]
      summary: Admin-only test case update
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestCaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCase"
    delete:
      operationId: deleteTestCase
      tags: [test-cases]
      summary: Delete a test case (admin only)
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
...
  /users/access-requests/{id}/approve:
    post:
      operationId: approveAccessRequest
      tags: [users]
      summary: Approve an access request
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessRequest"
  /users/access-requests/{id}/reject:
    post:
      operationId: rejectAccessRequest
      tags: [users]
      summary: Reject an access request with a reason
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectionInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessRequest"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Disable, enable, or change a user's role
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /profile/password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change current password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "204":
          description: Password changed
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
    Search:
      name: search
      in: query
      schema: { type: string }
    ModuleId:
      name: moduleId
      in: query
      schema: { type: integer }
    Result:
      name: result
      in: query
      schema:
        $ref: "#/components/schemas/TestResult"
    PerformedBy:
      name: performedBy
      in: query
      schema: { type: integer }
    From:
      name: from
      in: query
      schema: { type: string, format: date }
    To:
      name: to
      in: query
      schema: { type: string, format: date }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  schemas:
...
    User:
      type: object
      required: [id, fullName, email, role, accountStatus, createdAt]
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        accountStatus: { $ref: "#/components/schemas/AccountStatus" }
        createdAt: { type: string, format: date-time }
    Session:
      type: object
      required: [authenticated]
      properties:
        authenticated: { type: boolean }
        user:
          anyOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
    Role:
      type: string
      enum: [ADMIN, USER]
    AccountStatus:
      type: string
      enum: [PENDING, ACTIVE, REJECTED, DISABLED]
    TestResult:
      type: string
      enum: [PASS, FAIL, BLOCKED, NOT_TESTED]
...
    Module:
      type: object
      required: [id, name, code, status, testCaseCount, passCount, failCount, blockedCount]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        description: { type: ["string", "null"] }
        testCaseCount: { type: integer }
        passCount: { type: integer }
        failCount: { type: integer }
        blockedCount: { type: integer }
    TestCase:
      type: object
      required: [id, testCaseNumber, moduleId, moduleName, testCaseTag, testDate, description, expectedResult, actualResult, testResult, passedOn, performedBy, updatedAt, attachmentCount]
      properties:
        id: { type: integer }
        testCaseNumber: { type: string }
        moduleId: { type: integer }
        moduleName: { type: string }
        testCaseTag: { type: string }
        testDate: { type: string, format: date-time }
        description: { type: string }
        expectedResult: { type: string }
        actualResult: { type: string }
        testResult: { $ref: "#/components/schemas/TestResult" }
        passedOn: { type: ["string", "null"], format: date-time }
        performedBy: { type: string }
        performedByUserId: { type: integer }
        createdByUserId: { type: integer }
        createdBy: { type: string }
        updatedBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        attachmentCount: { type: integer }
    TestCaseDetail:
      allOf:
        - $ref: "#/components/schemas/TestCase"
        - type: object
          required: [attachments, history]
          properties:
            attachments:
              type: array
              items: { $ref: "#/components/schemas/Attachment" }
            history:
              type: array
              items: { $ref: "#/components/schemas/HistoryEntry" }
...
    TestCasePage:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TestCase" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    AccessRequest:
      type: object
      required: [id, userId, fullName, email, requestedRole, status, requestedAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        fullName: { type: string }
        email: { type: string, format: email }
        requestedRole: { $ref: "#/components/schemas/Role" }
        status: { type: string, enum: [PENDING, APPROVED, REJECTED] }
        rejectionReason: { type: ["string", "null"] }
        requestedAt: { type: string, format: date-time }
        reviewedAt: { type: ["string", "null"], format: date-time }
        reviewedBy: { type: ["string", "null"] }
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    AccessRequestInput:
      type: object
      required: [fullName, email, password, confirmPassword, requestedRole]
      properties:
        fullName: { type: string, minLength: 2 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        confirmPassword: { type: string, minLength: 8 }
        requestedRole: { $ref: "#/components/schemas/Role" }
    RejectionInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 3 }
    TestCaseInput:
      type: object
      required: [moduleId, testCaseTag, description, expectedResult, actualResult, testResult]
      properties:
        moduleId: { type: integer }
        testCaseTag: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        expectedResult: { type: string, minLength: 1 }
        actualResult: { type: string }
        testResult: { $ref: "#/components/schemas/TestResult" }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/AttachmentInput" }
    AttachmentInput:
      type: object
      required: [type, sourceType, url]
      properties:
        type: { type: string, enum: [IMAGE, VIDEO] }
        sourceType: { type: string, enum: [UPLOAD, EXTERNAL_LINK] }
        url: { type: string, format: uri }
        fileName: { type: string }
        mimeType: { type: string }
    TestCaseUpdate:
      type: object
      properties:
        moduleId: { type: integer }
        testCaseTag: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        expectedResult: { type: string, minLength: 1 }
        actualResult: { type: string }
        testResult: { $ref: "#/components/schemas/TestResult" }
        testDate: { type: string, format: date-time }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/AttachmentInput" }
    ModuleInput:
      type: object
      required: [name, code]
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1 }
        description: { type: string }
    ModuleUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1 }
        description: { type: string }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
    UserUpdate:
      type: object
      properties:
        role: { $ref: "#/components/schemas/Role" }
        accountStatus: { type: string, enum: [ACTIVE, DISABLED] }
    PasswordChangeInput:
      type: object
      required: [currentPassword, newPassword, confirmPassword]
      properties:
        currentPassword: { type: string, minLength: 8 }
        newPassword: { type: string, minLength: 8 }
        confirmPassword: { type: string, minLength: 8 }